chore: code cleanup
TCO is particularly inconvenient to implement for closures without
causing a leak and it's postponed for now.

+++ lifp/virtual_machine.h
@@ -6,8 +6,7 @@
 typedef struct environment_t {
   struct environment_t *parent;
   value_map_t values;
-
-  int refcount;
+  size_t refcount;
 } environment_t;
 
 typedef struct {
